abstract out a generic nether lily
